@@ -3,16 +3,16 @@ package com.vkram2711.usadba.activity
 import android.Manifest
 import android.content.Intent
 import android.os.Build
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import android.support.annotation.RequiresApi
-import android.support.v4.app.ActivityCompat
 import android.text.Editable
 import android.text.TextWatcher
 import android.view.View
 import android.widget.AdapterView
 import com.vkram2711.usadba.R
 import android.widget.ArrayAdapter
+import androidx.annotation.RequiresApi
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
 import com.vkram2711.usadba.Constants
 import com.vkram2711.usadba.callback.OnDataReceivedCallback
 import com.vkram2711.usadba.utils.DatabaseUtils
